pyproject.toml
[build-system]
requires = ["hatchling>=1.25"]
build-backend = "hatchling.build"

[project]
name = "readarr-to-play-books"
version = "0.1.0"
description = "Upload Readarr-imported books to Google Play Books via Playwright."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
  { name = "Community contributors" }
]
dependencies = [
  "playwright>=1.42",
  "python-dotenv>=1.0"
]

[project.optional-dependencies]
dev = [
  "ruff>=0.5",
  "black>=24.0",
  "mypy>=1.10",
  "pre-commit>=3.6"
]

[tool.ruff]
line-length = 100
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "I", "B", "W", "UP", "SIM"]
ignore = []

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.mypy]
python_version = "3.10"
warn_unused_configs = true
check_untyped_defs = true
disallow_untyped_defs = false
ignore_missing_imports = true
strict_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
no_implicit_optional = true

[project.urls]
Homepage = "https://github.com/your-org/readarr-to-play-books"
Issues = "https://github.com/your-org/readarr-to-play-books/issues"

[project.scripts]
readarr-to-play-books = "readarr_to_play_books.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/readarr_to_play_books"]


